+++ pick6/templates/index.html
@@ -1,40 +1,66 @@
 <!DOCTYPE html>
 <html lang="en">
 
-  <link rel="stylesheet" type="text/css" href="static/style.css">
+<link rel="stylesheet" type="text/css" href="static/style.css">
 
-  <head>
-    {% block head %}
+<head>
+  {% block head %}
 
-    <title>Mowtown Beatdown</title>
+  <title>Mowtown Beatdown</title>
 
-    <h1>
-      <b>
-        <p align="center">
-          <span style="color.red">Motown Beatdown</span>
-        </p>
-      </b>
-    </h1>
+  <h1>
+    <b>
+      <p align="center">
+        <span style="color.red">Motown Beatdown</span>
+      </p>
+    </b>
+  </h1>
 
-    <hr width="100%" height="25px" color="black">
+  <hr width="100%" height="25px" color="black">
 
-    {% endblock %}
-  </head>
+  {% endblock %}
+</head>
 
-  <body>
-    {% block body %}
+<body>
+  {% block body %}
 
-    <p align="center">
-      <img width="350px" height="300px", src="static/images/wagonwheel.jpg">
-    </p>
-    <p align="center">
-      <a href="https://www.officefootballpool.com/picks" target=_blank">WanPing Zhang Zen</a>
-    </p>
-    <p align="center">
-    <a href="http://127.0.0.1:5000/admin/" color = "black" target=_blank">Make Your Pick</a>
-    </p>
+  <p align="center">
+    <img width="350px" height="300px" , src="static/images/wagonwheel.jpg">
+  </p>
+  <p align="center">
+    <a href="https://www.officefootballpool.com/picks" target=_blank">WanPing Zhang Zen</a>
+  </p>
+  <p align="center">
+    <a href="http://127.0.0.1:5000/admin/" color="black" target=_blank">Make Your Pick</a>
+  </p>
 
-    {% endblock %}
-  </body>
+  <h1 align="center">
+    Last 6 picks
+  </h1>
+
+  <table width="1000" spacing="20" align="center">
+
+    <tr text-align="center">
+      <th>Date Entered</th>
+      <th>User</th>
+      <th>Team</th>
+      <th>Bet Type</th>
+      <th>Line</th>
+      <th>Result</th>
+    </tr>
+    {% for team in latest_team_list %}
+    <tr width="1000" spacing="20" align="center" text-align="center">
+      <td>{{ team.pub_date }}</td>
+      <td>{{ team.user_name }}</td>
+      <td>{{ team.team_text }}</td>
+      <td>{{ team.line_text }}</td>
+      <td>{{ team.line_pts }}</td>
+      <td>{{ team.result }}</td>
+    </tr>
+    {% endfor %}
+  </table>
+
+  {% endblock %}
+</body>
 
 </html>
